Keypad, Addressing, Protothreads
Known as "oled-temp-meter-by-address.ino". This version was finished 10/12/2020. It adds in the use of a keypad into the project, as well as switching from using the index's of the sensors to using them by their addresses. Various libraries and variables were added and methods were changed. Protothreads ("pt.h") were added to increase the response speed of the keypad, as there was often a delay or failure to react to human input of button presses. The main methods were divided up between what the oled displayed/keypad input and the temperature according to the sensors to better work with Protothreads.