Multiple Oled Displays
Known as "Chiller_Control_System.v2.ino". Finished on 12/13.2020. Deleted commented out code that was supposed to ask user to input set-point temperatures for sensors at the start of the program. Instead of a single oled display, multiple are now used. Methods and variables were modified to reflect this change.